refactor(analysis): rename language folder (#3994)
