+++ analysis/import/CoverageImporter/src/test/kotlin/ParserDialogTest.kt
@@ -5,7 +5,7 @@ import com.varabyte.kotter.runtime.terminal.inmemory.type
 import com.varabyte.kotterx.test.foundation.testSession
 import de.maibornwolff.codecharta.importer.coverage.CoverageImporter
 import de.maibornwolff.codecharta.importer.coverage.ParserDialog
-import de.maibornwolff.codecharta.importer.coverage.languages.Language
+import de.maibornwolff.codecharta.importer.coverage.Language
 import io.mockk.every
 import io.mockk.mockkObject
 import org.assertj.core.api.Assertions.assertThat
